moved autonomous code to autonomous package
