+++ TeamCode/src/main/java/org/firstinspires/ftc/teamcode/autonomous/environment/EnvironmentMap.java
@@ -1,11 +1,11 @@
-package org.firstinspires.ftc.teamcode.environment;
+package org.firstinspires.ftc.teamcode.autonomous.environment;
 
 import java.util.ArrayList;
 
-import org.firstinspires.ftc.teamcode.environment.NoActiveEntityException;
-import org.firstinspires.ftc.teamcode.environment.Entity;
-import org.firstinspires.ftc.teamcode.environment.EntityType;
-import org.firstinspires.ftc.teamcode.environment.RobotEntity;
+import org.firstinspires.ftc.teamcode.autonomous.environment.NoActiveEntityException;
+import org.firstinspires.ftc.teamcode.autonomous.environment.Entity;
+import org.firstinspires.ftc.teamcode.autonomous.environment.EntityType;
+import org.firstinspires.ftc.teamcode.autonomous.environment.RobotEntity;
 
 /** A wrapper for a list of entities. Provides a useful API for code abstraction, along
  * wth making the code easier for less advanced developers. */
