+++ TeamCode/src/main/java/org/firstinspires/ftc/teamcode/autonomous/environment/EnvironmentMapError.java
@@ -1,4 +1,4 @@
-package org.firstinspires.ftc.teamcode.environment;
+package org.firstinspires.ftc.teamcode.autonomous.environment;
 
 class NoActiveEntityException extends Exception {
     public NoActiveEntityException() { }
