+++ TeamCode/src/main/java/org/firstinspires/ftc/teamcode/autonomous/environment/GenericEntity.java
@@ -1,7 +1,7 @@
-package org.firstinspires.ftc.teamcode.environment;
+package org.firstinspires.ftc.teamcode.autonomous.environment;
 
-import org.firstinspires.ftc.teamcode.environment.Entity;
-import org.firstinspires.ftc.teamcode.environment.EntityType;
+import org.firstinspires.ftc.teamcode.autonomous.environment.Entity;
+import org.firstinspires.ftc.teamcode.autonomous.environment.EntityType;
 
 /** A generic entity. Used mostly for testing, but should be in
  * the main source tree as it's not actually a test itself. */
