+++ TeamCode/src/main/java/org/firstinspires/ftc/teamcode/autonomous/sensor/Sensor.java
@@ -1,4 +1,4 @@
-package org.firstinspires.ftc.teamcode.sensor;
+package org.firstinspires.ftc.teamcode.autonomous.sensor;
 
 /** An interface that all sensors should implement. This is mainly 
  * just for convience */
