+++ TeamCode/src/test/java/org/firstinspires/ftc/teamcode/autonomous/environment/EntityTypeTest.java
@@ -7,7 +7,7 @@ import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.RepeatedTest;
 import org.junit.jupiter.api.Test;
 
-import org.firstinspires.ftc.teamcode.environment.EntityType;
+import org.firstinspires.ftc.teamcode.autonomous.environment.EntityType;
 
 class EntityTypeTest {
     @Test
