+++ TeamCode/src/test/java/org/firstinspires/ftc/teamcode/autonomous/environment/RobotEntityTest.java
@@ -7,8 +7,8 @@ import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.RepeatedTest;
 import org.junit.jupiter.api.Test;
 
-import org.firstinspires.ftc.teamcode.environment.RobotEntity;
-import org.firstinspires.ftc.teamcode.environment.EntityType;
+import org.firstinspires.ftc.teamcode.autonomous.environment.RobotEntity;
+import org.firstinspires.ftc.teamcode.autonomous.environment.EntityType;
 
 class RobotEntityTest {
     RobotEntity entity;
